Update structure, put routes into their own files
- Put stub tests based on routes into separate files
- Rename route methods
- Update trailing slash in tests because of the APIRouter
